feat: multples imagenes, delete media available
- Se ha implementado la capacidad de subir multiples imagenes
- Se ha implementado endpoint de borrado

+++ src/resources/medias/medias.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { Inject, Injectable, Logger, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { CreateMediaDto } from "./dto/create-media.dto";
@@ -6,6 +6,9 @@ import { UpdateMediaDto } from "./dto/update-media.dto";
 import { Media } from "./entities/media.entity";
 import { File } from "@nest-lab/fastify-multer";
 import { ConfigService } from "@nestjs/config";
+import { CreateMediasDto } from "./dto/create-medias.dto";
+import path from "path";
+import { promises as fs } from "fs";
 
 @Injectable()
 export class MediasService {
@@ -16,6 +19,7 @@ export class MediasService {
 		@InjectRepository(Media)
 		private readonly mediaRepository: Repository<Media>,
 		private readonly configService: ConfigService,
+		@Inject(Logger) private readonly logger: Logger,
 	) {
 		const getLocationConfig = (key: string) => {
 			return this.configService.getOrThrow<string>(key).replace("./", "");
@@ -46,6 +50,12 @@ export class MediasService {
 	async remove(id: string) {
 		const media = await this.findOne(id);
 
+		if (!media) {
+			throw new NotFoundException("Media does not exist!");
+		}
+
+		await this.deleteFile(media.url);
+
 		return await this.mediaRepository.remove(media);
 	}
 
@@ -64,7 +74,29 @@ export class MediasService {
 		return await this.mediaRepository.save(createMediaDto);
 	}
 
-	async createImages(filesImage: Array<File>) {
-		console.log(filesImage.map((file) => file.filename));
+	async createImages(
+		createMediasDto: CreateMediasDto,
+		filesImage: Array<File>,
+	) {
+		const createMediaDtoList: CreateMediaDto[] = createMediasDto.alt.map(
+			(alt, index) => {
+				return new CreateMediaDto(
+					`${this.upload_location_images}/${filesImage[index].filename}`,
+					alt,
+				);
+			},
+		);
+		return await this.mediaRepository.save(createMediaDtoList);
+	}
+
+	private async deleteFile(url: string) {
+		const filePath = path.join(__dirname, "../../../", url);
+		try {
+			await fs.unlink(filePath);
+			this.logger.log(`File ${filePath.split("\\").at(-1)} deleted successfully`);
+		} catch (err) {
+			this.logger.error(`Error deleting file ${filePath}`, err);
+			throw err;
+		}
 	}
 }
